Automatically backed up by Learn

+++ index.js
@@ -33,7 +33,7 @@ function makeInt(n) {
 }
 
 function preserveDecimal(n) {
-  parseInt(n)
+  parseFloat(n)
   return n
 }
